added new page for DB stat calls
